pyproject.toml
[tool.poetry]
name = "directors-cut-rpg-discord-bot"
version = "0.1.0"
description = "A Discord dice bot for tabletop RPGs using the Director's Cut system, such as Outgunned or Household"
authors = ["Mirko Froehlich <682+digitalhobbit@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

packages = [
    { include = "bot" },
]

[tool.poetry.dependencies]
python = "^3.12"
"discord.py" = "^2.4.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
